api/main.py
"""FastAPI main application with route declarations."""

from fastapi import FastAPI, File, UploadFile, HTTPException, Depends
from fastapi.responses import JSONResponse
from typing import List, Optional
from pydantic import BaseModel
import os
from dotenv import load_dotenv

from api.storage.vector_store import VectorStore
from api.storage.file_store import FileStore
from api.services.ingestion import IngestionService
from api.services.retrieval import RetrievalService
from api.services.llm import LLMService

load_dotenv()

app = FastAPI(
    title="IN'SMART API",
    description="RAG-based educational assistant API",
    version="0.1.0",
)

# Initialize services (to be implemented)
vector_store: Optional[VectorStore] = None
file_store: Optional[FileStore] = None
ingestion_service: Optional[IngestionService] = None
retrieval_service: Optional[RetrievalService] = None
llm_service: Optional[LLMService] = None


@app.on_event("startup")
async def startup_event() -> None:
    """Initialize services on application startup.
    
    This method should:
    - Initialize VectorStore connection to Weaviate
    - Initialize FileStore connection to MinIO
    - Create IngestionService instance
    - Create RetrievalService instance
    - Create LLMService instance
    """
    pass


@app.on_event("shutdown")
async def shutdown_event() -> None:
    """Clean up resources on application shutdown.
    
    This method should:
    - Close database connections
    - Clean up any open resources
    """
    pass


# Pydantic models for request/response
class ChatMessage(BaseModel):
    """Chat message request model."""
    message: str
    conversation_id: Optional[str] = None


class ChatResponse(BaseModel):
    """Chat response model."""
    response: str
    sources: List[str]
    conversation_id: str


class DocumentMetadata(BaseModel):
    """Document metadata model."""
    document_id: str
    filename: str
    file_type: str
    upload_date: str
    status: str


class HealthResponse(BaseModel):
    """Health check response model."""
    status: str
    services: dict


@app.get("/health", response_model=HealthResponse)
async def health_check() -> HealthResponse:
    """Check API health and service connectivity.
    
    Returns:
        HealthResponse: Status of API and all connected services
            (Weaviate, MinIO, Mistral API)
    """
    pass


@app.post("/api/documents/upload")
async def upload_document(
    file: UploadFile = File(...),
    document_type: Optional[str] = None
) -> JSONResponse:
    """Upload a document for processing.
    
    Args:
        file: The file to upload (PDF, audio, or image)
        document_type: Optional type hint (pdf, audio, image)
    
    Returns:
        JSONResponse: Document ID and upload status
    
    Raises:
        HTTPException: If upload fails or file type is unsupported
    """
    pass


@app.get("/api/documents", response_model=List[DocumentMetadata])
async def list_documents() -> List[DocumentMetadata]:
    """List all uploaded documents.
    
    Returns:
        List[DocumentMetadata]: List of all documents with metadata
    """
    pass


@app.get("/api/documents/{document_id}", response_model=DocumentMetadata)
async def get_document(document_id: str) -> DocumentMetadata:
    """Get document metadata by ID.
    
    Args:
        document_id: Unique document identifier
    
    Returns:
        DocumentMetadata: Document information
    
    Raises:
        HTTPException: If document not found
    """
    pass


@app.delete("/api/documents/{document_id}")
async def delete_document(document_id: str) -> JSONResponse:
    """Delete a document and all associated chunks.
    
    Args:
        document_id: Unique document identifier
    
    Returns:
        JSONResponse: Deletion status
    
    Raises:
        HTTPException: If document not found or deletion fails
    """
    pass


@app.post("/api/ingest/{document_id}")
async def ingest_document(document_id: str) -> JSONResponse:
    """Process and ingest a document into the vector database.
    
    This endpoint:
    1. Retrieves the document from file storage
    2. Processes it (STT/OCR/text extraction)
    3. Chunks the text
    4. Generates embeddings
    5. Stores chunks in vector database
    
    Args:
        document_id: Unique document identifier
    
    Returns:
        JSONResponse: Ingestion status and chunk count
    
    Raises:
        HTTPException: If document not found or ingestion fails
    """
    pass


@app.post("/api/chat", response_model=ChatResponse)
async def chat(message: ChatMessage) -> ChatResponse:
    """Send a message and get AI response with RAG.
    
    This endpoint:
    1. Retrieves relevant document chunks using semantic search
    2. Assembles context from retrieved chunks
    3. Generates response using LLM with context
    4. Returns response with source citations
    
    Args:
        message: Chat message request with optional conversation_id
    
    Returns:
        ChatResponse: AI response with source document references
    
    Raises:
        HTTPException: If retrieval or generation fails
    """
    pass


@app.get("/api/chat/history")
async def get_chat_history(
    conversation_id: str,
    limit: int = 50
) -> JSONResponse:
    """Get chat history for a conversation.
    
    Args:
        conversation_id: Unique conversation identifier
        limit: Maximum number of messages to return
    
    Returns:
        JSONResponse: List of messages in the conversation
    
    Raises:
        HTTPException: If conversation not found
    """
    pass


if __name__ == "__main__":
    import uvicorn
    uvicorn.run(
        "main:app",
        host=os.getenv("API_HOST", "0.0.0.0"),
        port=int(os.getenv("API_PORT", 8000)),
        reload=True
    )
