api/storage/vector_store.py
"""Vector store connector for Weaviate."""

from typing import List, Dict, Optional, Any
import weaviate
from weaviate.classes.config import Configure, Property, DataType


class VectorStore:
    """Weaviate client wrapper for vector storage operations."""

    def __init__(self, url: str, class_name: str = "DocumentChunk") -> None:
        """Initialize Weaviate client.
        
        Args:
            url: Weaviate server URL
            class_name: Name of the Weaviate class for document chunks
        """
        self.url = url
        self.class_name = class_name
        self.client: Optional[weaviate.WeaviateClient] = None

    def connect(self) -> None:
        """Establish connection to Weaviate server.
        
        Raises:
            ConnectionError: If connection to Weaviate fails
        """
        pass

    def disconnect(self) -> None:
        """Close connection to Weaviate server."""
        pass

    def create_schema(self) -> None:
        """Create Weaviate schema for document chunks.
        
        Schema should include:
        - text: The chunk text content
        - document_id: Reference to source document
        - chunk_index: Position of chunk in document
        - metadata: Additional metadata (page number, timestamp, etc.)
        
        Raises:
            ValueError: If schema creation fails
        """
        pass

    def delete_schema(self) -> None:
        """Delete the Weaviate schema and all data.
        
        Warning: This will delete all stored chunks.
        
        Raises:
            ValueError: If schema deletion fails
        """
        pass

    def insert_chunks(
        self,
        chunks: List[Dict[str, Any]],
        embeddings: List[List[float]]
    ) -> List[str]:
        """Insert document chunks with embeddings into Weaviate.
        
        Args:
            chunks: List of chunk dictionaries with text, document_id, chunk_index, metadata
            embeddings: List of embedding vectors for each chunk
        
        Returns:
            List[str]: List of Weaviate UUIDs for inserted chunks
        
        Raises:
            ValueError: If chunks and embeddings length mismatch or insertion fails
        """
        pass

    def search_similar(
        self,
        query_embedding: List[float],
        limit: int = 5,
        where_filter: Optional[Dict[str, Any]] = None
    ) -> List[Dict[str, Any]]:
        """Search for similar chunks using vector similarity.
        
        Args:
            query_embedding: Query vector embedding
            limit: Maximum number of results to return
            where_filter: Optional filter conditions (e.g., by document_id)
        
        Returns:
            List[Dict[str, Any]]: List of similar chunks with:
                - text: Chunk text
                - document_id: Source document ID
                - chunk_index: Chunk position
                - metadata: Additional metadata
                - distance: Similarity distance
                - uuid: Weaviate UUID
        
        Raises:
            ValueError: If search fails
        """
        pass

    def delete_by_document_id(self, document_id: str) -> int:
        """Delete all chunks associated with a document.
        
        Args:
            document_id: Document identifier
        
        Returns:
            int: Number of chunks deleted
        
        Raises:
            ValueError: If deletion fails
        """
        pass

    def get_chunk_by_id(self, chunk_id: str) -> Optional[Dict[str, Any]]:
        """Retrieve a chunk by its Weaviate UUID.
        
        Args:
            chunk_id: Weaviate UUID
        
        Returns:
            Optional[Dict[str, Any]]: Chunk data if found, None otherwise
        
        Raises:
            ValueError: If retrieval fails
        """
        pass

    def count_chunks(self, document_id: Optional[str] = None) -> int:
        """Count total chunks, optionally filtered by document.
        
        Args:
            document_id: Optional document ID to filter by
        
        Returns:
            int: Total number of chunks
        
        Raises:
            ValueError: If count query fails
        """
        pass

    def batch_insert(
        self,
        chunks: List[Dict[str, Any]],
        embeddings: List[List[float]],
        batch_size: int = 100
    ) -> List[str]:
        """Insert chunks in batches for better performance.
        
        Args:
            chunks: List of chunk dictionaries
            embeddings: List of embedding vectors
            batch_size: Number of chunks per batch
        
        Returns:
            List[str]: List of Weaviate UUIDs for all inserted chunks
        
        Raises:
            ValueError: If batch insertion fails
        """
        pass
